+++ docs/source/versions/changelog.rst
@@ -3,6 +3,8 @@ Change log
 
 **Version 2.5.0** (Month, Day, 2022)
 
+ * New method :meth:`plot <ketos.audio.waveform.plot>` for plotting multiple waveforms superimposed on one another.
+ * Modified :meth:`resample <ketos.audio.waveform.Waveform.resample>` to use librosa's resampling function.
  * Bug fix in :meth:`recover_waveform <ketos.audio.spectrogram.MagSpectrogram.recover_waveform>` (ensure that output waveform as correct length).
  * Bug fix in parsing module (ensure that audio representations are encoded correctly when passed as a nested dict).
  * Bug fix in :meth:`train_loop <ketos.neural_networks.dev_utils.nn_interface.train_loop>` (ensure `last_epoch_with_improvement` attr is always initialized)